test: move createEmailData tests to SubmissionEmailObj
