+++ src/app/modules/submission/email-submission/__tests__/email-submission.service.spec.ts
@@ -8,40 +8,28 @@ import { mocked } from 'ts-jest/utils'
 import { getEmailSubmissionModel } from 'src/app/models/submission.server.model'
 import { DatabaseError } from 'src/app/modules/core/core.errors'
 import MailService from 'src/app/services/mail/mail.service'
-import { types as basicTypes } from 'src/shared/resources/basic'
 import {
   AuthType,
   BasicField,
   IEmailFormSchema,
   IEmailSubmissionSchema,
   IPopulatedEmailForm,
-  MyInfoAttribute,
   SubmissionType,
 } from 'src/types'
 
-import {
-  generateSingleAnswerAutoreply,
-  generateSingleAnswerFormData,
-  generateSingleAnswerJson,
-} from 'tests/unit/backend/helpers/generate-email-data'
+import { generateSingleAnswerFormData } from 'tests/unit/backend/helpers/generate-email-data'
 import {
   generateNewAttachmentResponse,
-  generateNewCheckboxResponse,
   generateNewSingleAnswerResponse,
-  generateNewTableResponse,
 } from 'tests/unit/backend/helpers/generate-form-data'
 
 import { SendAdminEmailError } from '../../submission.errors'
 import { ProcessedSingleAnswerResponse } from '../../submission.types'
 import {
-  ATTACHMENT_PREFIX,
   DIGEST_TYPE,
   HASH_ITERATIONS,
   KEY_LENGTH,
-  MYINFO_PREFIX,
   SALT_LENGTH,
-  TABLE_PREFIX,
-  VERIFIED_PREFIX,
 } from '../email-submission.constants'
 import {
   AttachmentTooLargeError,
@@ -61,356 +49,7 @@ const MOCK_HASH = Buffer.from('mockHash')
 
 const EmailSubmissionModel = getEmailSubmissionModel(mongoose)
 
-const ALL_SINGLE_SUBMITTED_RESPONSES = basicTypes
-  // Attachments are special cases, requiring filename and content
-  // Section fields are not submitted
-  .filter(
-    (t) =>
-      !t.answerArray &&
-      ![BasicField.Attachment, BasicField.Section].includes(t.name),
-  )
-  .map((t) => generateNewSingleAnswerResponse(t.name))
-
 describe('email-submission.service', () => {
-  describe('createEmailData', () => {
-    it('should return email data correctly for all single answer field types', () => {
-      const emailData = EmailSubmissionService.createEmailData(
-        ALL_SINGLE_SUBMITTED_RESPONSES,
-        new Set(),
-      )
-      const expectedAutoReplyData = ALL_SINGLE_SUBMITTED_RESPONSES.map(
-        generateSingleAnswerAutoreply,
-      )
-      const expectedDataCollationData = ALL_SINGLE_SUBMITTED_RESPONSES.map(
-        generateSingleAnswerJson,
-      )
-      const expectedFormData = ALL_SINGLE_SUBMITTED_RESPONSES.map(
-        generateSingleAnswerFormData,
-      )
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should exclude section fields from JSON data', () => {
-      const response = generateNewSingleAnswerResponse(BasicField.Section)
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-      expect(emailData.dataCollationData).toEqual([])
-      expect(emailData.autoReplyData).toEqual([
-        generateSingleAnswerAutoreply(response),
-      ])
-      expect(emailData.formData).toEqual([
-        generateSingleAnswerFormData(response),
-      ])
-    })
-
-    it('should exclude non-visible fields from autoreply data', () => {
-      const response = generateNewSingleAnswerResponse(BasicField.ShortText, {
-        isVisible: false,
-      })
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      expect(emailData.dataCollationData).toEqual([
-        generateSingleAnswerJson(response),
-      ])
-      expect(emailData.autoReplyData).toEqual([])
-      expect(emailData.formData).toEqual([
-        generateSingleAnswerFormData(response),
-      ])
-    })
-
-    it('should generate table answers with [table] prefix in form and JSON data', () => {
-      const response = generateNewTableResponse()
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-      const firstRow = response.answerArray[0].join(',')
-      const secondRow = response.answerArray[1].join(',')
-
-      const expectedDataCollationData = [
-        { question: `${TABLE_PREFIX}${question}`, answer: firstRow },
-        { question: `${TABLE_PREFIX}${question}`, answer: secondRow },
-      ]
-
-      const expectedAutoReplyData = [
-        { question, answerTemplate: [firstRow] },
-        { question, answerTemplate: [secondRow] },
-      ]
-
-      const expectedFormData = [
-        {
-          question: `${TABLE_PREFIX}${question}`,
-          answer: firstRow,
-          answerTemplate: [firstRow],
-          fieldType: BasicField.Table,
-        },
-        {
-          question: `${TABLE_PREFIX}${question}`,
-          answer: secondRow,
-          answerTemplate: [secondRow],
-          fieldType: BasicField.Table,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should generate checkbox answers correctly', () => {
-      const response = generateNewCheckboxResponse()
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-      const answer = response.answerArray.join(', ')
-
-      const expectedDataCollationData = [{ question, answer }]
-      const expectedAutoReplyData = [{ question, answerTemplate: [answer] }]
-      const expectedFormData = [
-        {
-          question,
-          answer,
-          answerTemplate: [answer],
-          fieldType: BasicField.Checkbox,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should generate attachment answers with [attachment] prefix in form and JSON data', () => {
-      const response = generateNewAttachmentResponse()
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-      const answer = response.answer
-
-      const expectedDataCollationData = [
-        { question: `${ATTACHMENT_PREFIX}${question}`, answer },
-      ]
-      const expectedAutoReplyData = [{ question, answerTemplate: [answer] }]
-      const expectedFormData = [
-        {
-          question: `${ATTACHMENT_PREFIX}${question}`,
-          answer,
-          answerTemplate: [answer],
-          fieldType: BasicField.Attachment,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should split single answer fields by newline', () => {
-      const answer = 'first line\nsecond line'
-      const response = generateNewSingleAnswerResponse(BasicField.ShortText, {
-        answer,
-      })
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-
-      const expectedDataCollationData = [{ question, answer }]
-      const expectedAutoReplyData = [
-        { question, answerTemplate: answer.split('\n') },
-      ]
-      const expectedFormData = [
-        {
-          question,
-          answer,
-          answerTemplate: answer.split('\n'),
-          fieldType: BasicField.ShortText,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should split table answers by newline', () => {
-      const answerArray = [['firstLine\nsecondLine', 'thirdLine\nfourthLine']]
-      const response = generateNewTableResponse({ answerArray })
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-      const answer = answerArray[0].join(',')
-
-      const expectedDataCollationData = [
-        { question: `${TABLE_PREFIX}${question}`, answer },
-      ]
-      const expectedAutoReplyData = [
-        { question, answerTemplate: answer.split('\n') },
-      ]
-      const expectedFormData = [
-        {
-          question: `${TABLE_PREFIX}${question}`,
-          answer,
-          answerTemplate: answer.split('\n'),
-          fieldType: BasicField.Table,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should split checkbox answers by newline', () => {
-      const answerArray = ['firstLine\nsecondLine', 'thirdLine\nfourtLine']
-      const response = generateNewCheckboxResponse({ answerArray })
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-      const answer = answerArray.join(', ')
-
-      const expectedDataCollationData = [{ question, answer }]
-      const expectedAutoReplyData = [
-        { question, answerTemplate: answer.split('\n') },
-      ]
-      const expectedFormData = [
-        {
-          question,
-          answer,
-          answerTemplate: answer.split('\n'),
-          fieldType: BasicField.Checkbox,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should prefix verified fields with [verified] only in form data', () => {
-      const response = generateNewSingleAnswerResponse(BasicField.Email, {
-        isUserVerified: true,
-      })
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [response],
-        new Set(),
-      )
-
-      const question = response.question
-      const answer = response.answer
-
-      const expectedDataCollationData = [{ question, answer }]
-      const expectedAutoReplyData = [{ question, answerTemplate: [answer] }]
-      const expectedFormData = [
-        {
-          question: `${VERIFIED_PREFIX}${question}`,
-          answer,
-          answerTemplate: [answer],
-          fieldType: BasicField.Email,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-
-    it('should prefix MyInfo-verified fields with [MyInfo] only in form data', () => {
-      // MyInfo-verified
-      const nameResponse = generateNewSingleAnswerResponse(
-        BasicField.ShortText,
-        {
-          myInfo: { attr: MyInfoAttribute.Name },
-          answer: 'name',
-        },
-      )
-
-      // MyInfo field but not MyInfo-verified
-      const vehicleResponse = generateNewSingleAnswerResponse(
-        BasicField.ShortText,
-        {
-          myInfo: { attr: MyInfoAttribute.VehicleNo },
-          answer: 'vehiclenumber',
-        },
-      )
-
-      const emailData = EmailSubmissionService.createEmailData(
-        [nameResponse, vehicleResponse],
-        new Set([nameResponse._id]),
-      )
-
-      const expectedDataCollationData = [
-        { question: nameResponse.question, answer: nameResponse.answer },
-        {
-          question: vehicleResponse.question,
-          answer: vehicleResponse.answer,
-        },
-      ]
-      const expectedAutoReplyData = [
-        {
-          question: nameResponse.question,
-          answerTemplate: [nameResponse.answer],
-        },
-        {
-          question: vehicleResponse.question,
-          answerTemplate: [vehicleResponse.answer],
-        },
-      ]
-      const expectedFormData = [
-        {
-          // Prefixed because its ID was in the Set
-          question: `${MYINFO_PREFIX}${nameResponse.question}`,
-          answer: nameResponse.answer,
-          answerTemplate: [nameResponse.answer],
-          fieldType: BasicField.ShortText,
-        },
-        {
-          // Not prefixed because ID not in Set
-          question: vehicleResponse.question,
-          answer: vehicleResponse.answer,
-          answerTemplate: [vehicleResponse.answer],
-          fieldType: BasicField.ShortText,
-        },
-      ]
-
-      expect(emailData.dataCollationData).toEqual(expectedDataCollationData)
-      expect(emailData.autoReplyData).toEqual(expectedAutoReplyData)
-      expect(emailData.formData).toEqual(expectedFormData)
-    })
-  })
-
   describe('validateAttachments', () => {
     it('should reject submissions when attachments are more than 7MB', async () => {
       const processedResponse1 = generateNewAttachmentResponse({
